requirements.txt
pypdf
google-generativeai
python-dotenv
httpx
